refactor: base model in common app
