+++ bc_obps/common/constants.py
@@ -1,0 +1,9 @@
+# Used to exclude audit fields from the schema
+AUDIT_FIELDS = [
+    "created_at",
+    "created_by",
+    "updated_at",
+    "updated_by",
+    "archived_at",
+    "archived_by",
+]
