+++ bc_obps/registration/models.py
@@ -2,6 +2,7 @@ from typing import List, Optional
 import re
 from uuid import UUID
 import uuid
+from common.models import BaseModel
 from django.db import models
 from phonenumber_field.modelfields import PhoneNumberField
 from localflavor.ca.models import CAPostalCodeField, CAProvinceField
@@ -11,7 +12,6 @@ from registration.constants import (
     CRA_BUSINESS_NUMBER_MESSAGE,
     BORO_ID_REGEX,
     USER_CACHE_PREFIX,
-    AUDIT_FIELDS,
 )
 from simple_history.models import HistoricalRecords
 from django.core.validators import RegexValidator, MaxValueValidator, MinValueValidator
@@ -19,28 +19,6 @@ from django.utils import timezone
 from django.core.exceptions import ValidationError
 from django.core.files.storage import default_storage
 from django.core.cache import cache
-
-
-class BaseModel(models.Model):
-    """
-    Abstract base class for all models in the app.
-    This class adds a save method that calls full_clean before saving.
-    """
-
-    class Meta:
-        abstract = True
-
-    def save(self, *args, **kwargs):
-        # if `update_fields` is passed, we only clean them otherwise we clean all fields(except audit fields)
-        # This is to optimize the performance of the save method by only validating the fields that are being updated
-        fields_to_update = kwargs.get('update_fields')
-        if fields_to_update:
-            self.full_clean(
-                exclude=[field.name for field in self._meta.fields if field.name not in fields_to_update]
-            )  # validate the model before saving
-        else:
-            self.full_clean(exclude=AUDIT_FIELDS)  # validate the model before saving
-        super().save(*args, **kwargs)
 
 
 class TimeStampedModelManager(models.Manager):
