+++ bc_obps/registration/schema/operator.py
@@ -1,8 +1,8 @@
 from typing import List, Optional
 from uuid import UUID
+from common.constants import AUDIT_FIELDS
 from ninja import Field, Schema
 from ninja import ModelSchema
-from registration.constants import AUDIT_FIELDS
 from registration.models import Operator
 from .parent_operator import ParentOperatorOut
