+++ bc_obps/registration/schema/user_operator.py
@@ -1,11 +1,12 @@
 from typing import List, Optional
 from uuid import UUID
+from common.constants import AUDIT_FIELDS
 from registration.schema.validators import validate_cra_business_number
 from registration.schema.operator import OperatorExternalDashboardUsersTileData
 from registration.schema.user import UserExternalDashboardUsersTileData
 from ninja import ModelSchema, Schema, Field
 from pydantic import field_validator
-from registration.constants import AUDIT_FIELDS, BC_CORPORATE_REGISTRY_REGEX
+from registration.constants import BC_CORPORATE_REGISTRY_REGEX
 from registration.models import BusinessStructure, Contact, UserOperator
 from .parent_operator import ParentOperatorIn, ParentOperatorOut
 from .business_structure import validate_business_structure
